feat: add initial interface and implementation for WASI NN
This commit begins the adaptation of the WASI NN interface to the new
WIT format, and adds a WIP implementation for the API using the Tract
ONNX implementation as the only inferencing engine (ideally, the same
implementation would also allow the execution of Tensorflow graphs,
potentially using the same Tract implementation).

This is currently a WIP, as there are no tests yet.

fix: remove doc tests from running for the NN implementation

chore: update wit-bindgen revision

Signed-off-by: Radu Matei <radu.matei@fermyon.com>