wasi-outbound-http: reorganize code
Move the WIT definitions to a dedicated crate. This crate can then be
imported by importers (Wasm modules) and exporters (Wasm hosts).

Prior to this change, all the WIT files were defined inside the root of
the `wasi-experimental-toolkit` crate. That worked fine for the
sub-crates and for the tests because they had direct access to the
files. However that didn't work when trying to consume these files
inside of standalone projects. That's because the WIT files were not
part of the modules checked out by cargo.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>